+++ ChangeLog.md
@@ -4,6 +4,7 @@ ADD:
  * Start adding support for MiniTV 1.44" 
  * MiniTV show time, date, status.
  * Fix total packets counter
+ * Oled_096 support (again ;) 
 
 -=# v0.0.4 (20260208) #=- 
 ADD: